+++ tests/sqllogictest/Cargo.toml
@@ -15,5 +15,9 @@ tempfile = "3"
 tokio = { version = "1", features = ["rt"] }
 
 [[test]]
-name = "sqllogictest"
+name = "disk"
+harness = false
+
+[[test]]
+name = "mem"
 harness = false
